refactor: MeetingResult 패키지 수정
