+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/meeting/MeetingAPI.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.infrastructure.web.meeting;
 
 import com.mobble.mobbleserver.application.meeting.command.request.CreateMeetingCommand;
 import com.mobble.mobbleserver.application.meeting.command.request.UpdateMeetingCommand;
-import com.mobble.mobbleserver.application.meeting.command.response.MeetingResult;
+import com.mobble.mobbleserver.application.meeting.result.MeetingResult;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingCreatePort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingDeletePort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingQueryPort;
