+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/meeting/dto/response/MeetingDetailResponseDto.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.infrastructure.web.meeting.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.mobble.mobbleserver.application.meeting.command.response.MeetingResult;
+import com.mobble.mobbleserver.application.meeting.result.MeetingResult;
 import com.mobble.mobbleserver.application.meetingMember.command.response.MeetingMemberInfoResult;
 import com.mobble.mobbleserver.domain.meeting.MeetingType;
